continue working on asynclogger for each trhead

+++ SGDEngine/Engine/H1PlatformUtil.h
@@ -4,16 +4,16 @@ namespace Util
 {
 		// bit operations
 		bool appBitScanReverse(uint32& Offset, uint32 Mask);
-		bool appBitScanReverse64(uint32& Offset, uint64 Mask);
+		bool appBitScanReverse64(uint64& Offset, uint64 Mask);
 		bool appBitScanForward(uint32& Offset, uint32 Mask);
-		bool appBitScanForward64(uint32& Offset, uint64 Mask);
+		bool appBitScanForward64(uint64& Offset, uint64 Mask);
 		// bit test and reset it to 0
 		bool appBitTestAndReset(uint32 Offset, uint32& Mask);
-		bool appBitTestAndReset64(uint32 Offset, uint64& Mask);
+		bool appBitTestAndReset64(uint64 Offset, uint64& Mask);
 		bool appBitTestAndSet(uint32 Offset, uint32& Mask);
-		bool appBitTestAndSet64(uint32 Offset, uint64& Mask);
+		bool appBitTestAndSet64(uint64 Offset, uint64& Mask);
 		bool appBitTest(uint32 Offset, uint32 Mask);
-		bool appBitTest64(uint32 Offset, uint64 Mask);
+		bool appBitTest64(uint64 Offset, uint64 Mask);
 
 		// output debug string
 		void appOutputDebugString(const char* String);
